+++ src/main/java/org/jax/mgi/mgd/api/entities/ReferenceCitationData.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class ReferenceCitationData extends EntityBase {
 	@Id
 	@Column(name="_refs_key")
-	private long _refs_key;
+	private int _refs_key;
 
 	@Column(name="short_citation")
 	private String short_citation;
